WIP - add some error handling to the lexer. This approach could be also used with the parser, although probably it'd be fine to explicitly pass the error context to the error-declaring combinators.
There should be a stack of error messages, so that setting a message in a combinator shouldn't overwrite existing messages if the combinator succeeds.

For some error cases, passing context from the parser to the lexer (e.g. '{' vs ';') would be useful as well. This is also needed to sanely support the embedded type syntax.